+++ tests/pytorch_tests/model_tests/feature_models/qat_test.py
@@ -30,9 +30,9 @@ from mct_quantizers.pytorch.quantizers import BasePyTorchInferableQuantizer
 from model_compression_toolkit.core.pytorch.pytorch_device_config import get_working_device
 from model_compression_toolkit.core.pytorch.utils import to_torch_tensor
 from model_compression_toolkit.qat.pytorch.quantizer.base_pytorch_qat_weight_quantizer import BasePytorchQATWeightTrainableQuantizer
-from model_compression_toolkit.trainable_infrastructure.pytorch.activation.base_activation_quantizer import \
+from model_compression_toolkit.trainable_infrastructure.pytorch.activation_quantizers.base_activation_quantizer import \
     BasePytorchActivationTrainableQuantizer
-from model_compression_toolkit.trainable_infrastructure.pytorch.activation.ste.symmetric_ste import \
+from model_compression_toolkit.trainable_infrastructure.pytorch.activation_quantizers.ste.symmetric_ste import \
     STESymmetricActivationTrainableQuantizer
 from model_compression_toolkit.target_platform_capabilities.tpc_models.imx500_tpc.latest import generate_pytorch_tpc
 from model_compression_toolkit.target_platform_capabilities.tpc_models.imx500_tpc.latest import \
